% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundle.R
\name{bundle}
\alias{bundle}
\alias{unbundle}
\title{Bundle And Unbundle Columns}
\usage{
bundle(data, ..., .key = "data")

unbundle(data, ...)
}
\arguments{
\item{data}{A data.frame.}

\item{...}{Selections of columns.}

\item{.key}{The name of the new column. If the dots have names, this will be ignored.}
}
\description{
Bundle And Unbundle Columns
}
