% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/named_capture_group.R
\name{str_extract_groups}
\alias{str_extract_groups}
\title{Extract Named Capture Groups}
\usage{
str_extract_groups(x, pattern)
}
\arguments{
\item{x}{A string.}

\item{pattern}{A regex.}
}
\description{
Extract Named Capture Groups
}
