% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiedr-package.R
\docType{package}
\name{tiedr-package}
\alias{tiedr}
\alias{tiedr-package}
\title{tiedr: Experimental Package For Multi-Gather And Multi-Spread}
\description{
Experimental Package For Multi-Gather And Multi-Spread.
}
\author{
\strong{Maintainer}: Hiroaki Yutani \email{yutani.ini@gamil.com}

}
\keyword{internal}
